+++ BaseClasses.py
@@ -148,7 +148,7 @@ class World(object):
             set_player_attr('door_self_loops', False)
             set_player_attr('door_type_mode', 'original')
             set_player_attr('trap_door_mode', 'optional')
-            set_player_attr('key_logic_algorithm', 'default')
+            set_player_attr('key_logic_algorithm', 'partial')
             set_player_attr('aga_randomness', True)
 
             set_player_attr('shopsanity', False)